+++ src/main/java/com/project/FoodsolomonBackend/bookmark/repository/BookmarkRepository.java
@@ -1,10 +1,10 @@
-package com.project.FoodsolomonBackend.repository;
+package com.project.FoodsolomonBackend.bookmark.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.project.FoodsolomonBackend.entity.Bookmark;
+import com.project.FoodsolomonBackend.bookmark.entity.Bookmark;
 
 public interface BookmarkRepository extends JpaRepository<Bookmark, Long>{
